server/pyproject.toml
[project]
name = "talky"
version = "0.1.0"
description = "Voice AI bot built with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[deepgram,google,kokoro,local,openrouter,runner,silero,webrtc,whisper]",
    "pyyaml",
    "loguru",
    "mlx-whisper>=0.4.3",
]
# Note: opencv-python excluded to avoid AVFoundation symbol conflicts with 'av' library
# Add opencv-python manually if you need computer vision features

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]
vision = [
    "opencv-python>=4.0.0",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
